UPDATE: add hover effect for search employee nav link
